Move timeline renderers into subdirectories
