+++ ui/src/components/Timeline2/TimelineChart/TimelineChart.tsx
@@ -2,8 +2,8 @@ import { FC, CSSProperties } from 'react';
 import { Classes } from '@blueprintjs/core';
 import c from 'classnames';
 import { differenceInDays, startOfMonth, endOfMonth } from 'date-fns';
-import { useTimelineKeyboardNavigation } from './util';
-import type { TimelineRendererProps } from './types';
+import { useTimelineKeyboardNavigation } from '../util';
+import type { TimelineRendererProps } from '../types';
 import TimelineChartGrid from './TimelineChartGrid';
 import TimelineChartLabels from './TimelineChartLabels';
 import TimelineChartItem from './TimelineChartItem';
