+++ ui/src/components/Timeline2/TimelineChart/index.ts
@@ -1,0 +1,1 @@
+export { default as TimelineChart } from './TimelineChart';
