+++ ui/src/components/Timeline2/TimelineList/TimelineList.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen, within } from 'testUtils';
 import userEvent from '@testing-library/user-event';
 import { Model, defaultModel } from '@alephdata/followthemoney';
-import { TimelineItem } from './util';
+import { TimelineItem } from '../util';
 import TimelineList from './TimelineList';
 
 const model = new Model(defaultModel);
