+++ ui/src/components/Timeline2/TimelineList/index.ts
@@ -1,0 +1,1 @@
+export { default as TimelineList } from './TimelineList';
